refactor(codegen): extract arithmetic helpers into arith.rs
Relocate the standalone integer/float arithmetic helpers out of the llvm.rs
god-module into a sibling arith module: overflow_intrinsic_name,
saturating_intrinsic_name, saturating_bound, lower_saturating_width_cast, and
the std.mem floor-intrinsic synthesis (FloorIntrinsic + emit_floor_intrinsic_body).
The per-instruction binary-op lowering stays in lower_instruction's match arms;
only the free helpers those arms call move.

Pure relocation following the coro.rs carve: arith imports the shared codegen
context from crate::llvm, and llvm.rs calls back via crate::arith::. Every
emitted .ll is byte-identical before and after (ll-corpus verify clean on all
fixtures, native and wasm32).
