+++ hew-codegen-rs/src/llvm.rs
@@ -12575,33 +12575,6 @@ fn emit_spawn_task_closure(
     Ok(())
 }
 
-fn overflow_intrinsic_name(op: IntArithOp, signed: IntSignedness) -> &'static str {
-    match (op, signed) {
-        (IntArithOp::Add, IntSignedness::Signed) => "llvm.sadd.with.overflow",
-        (IntArithOp::Add, IntSignedness::Unsigned) => "llvm.uadd.with.overflow",
-        (IntArithOp::Sub, IntSignedness::Signed) => "llvm.ssub.with.overflow",
-        (IntArithOp::Sub, IntSignedness::Unsigned) => "llvm.usub.with.overflow",
-        (IntArithOp::Mul, IntSignedness::Signed) => "llvm.smul.with.overflow",
-        (IntArithOp::Mul, IntSignedness::Unsigned) => "llvm.umul.with.overflow",
-    }
-}
-
-/// Return the `llvm.{s,u}{add,sub}.sat` base name for Add/Sub.
-/// LLVM does not provide `llvm.smul.sat`; callers must keep the
-/// `.with.overflow` + select path for Mul. Panics on Mul to enforce the
-/// call-site invariant (the `IntArithSaturating` arm only routes Add/Sub here).
-fn saturating_intrinsic_name(op: IntArithOp, signed: IntSignedness) -> &'static str {
-    match (op, signed) {
-        (IntArithOp::Add, IntSignedness::Signed) => "llvm.sadd.sat",
-        (IntArithOp::Add, IntSignedness::Unsigned) => "llvm.uadd.sat",
-        (IntArithOp::Sub, IntSignedness::Signed) => "llvm.ssub.sat",
-        (IntArithOp::Sub, IntSignedness::Unsigned) => "llvm.usub.sat",
-        (IntArithOp::Mul, _) => {
-            panic!("saturating_intrinsic_name: llvm.smul.sat does not exist; use overflow path for Mul")
-        }
-    }
-}
-
 fn validate_numeric_method_width(
     width: NumericWidth,
     int_ty: inkwell::types::IntType<'_>,
@@ -12616,57 +12589,6 @@ fn validate_numeric_method_width(
         NumericWidth::Pointer => Err(CodegenError::FailClosed(format!(
             "{construct} on isize/usize requires target pointer-width layout; no target layout was provided to this instruction"
         ))),
-    }
-}
-
-fn saturating_bound<'ctx>(
-    fn_ctx: &FnCtx<'_, 'ctx>,
-    op: IntArithOp,
-    signed: IntSignedness,
-    int_ty: inkwell::types::IntType<'ctx>,
-    lhs_v: IntValue<'ctx>,
-    rhs_v: IntValue<'ctx>,
-) -> CodegenResult<IntValue<'ctx>> {
-    let bits = int_ty.get_bit_width();
-    let zero = int_ty.const_zero();
-    match signed {
-        IntSignedness::Unsigned => Ok(match op {
-            IntArithOp::Sub => zero,
-            IntArithOp::Add | IntArithOp::Mul => int_ty.const_all_ones(),
-        }),
-        IntSignedness::Signed => {
-            let max = int_ty.const_int(((1u128 << (bits - 1)) - 1) as u64, false);
-            let min = int_ty.const_int((1u128 << (bits - 1)) as u64, false);
-            let choose_max = match op {
-                IntArithOp::Add => fn_ctx
-                    .builder
-                    .build_int_compare(IntPredicate::SGE, rhs_v, zero, "sat_add_positive")
-                    .llvm_ctx("saturating add sign compare")?,
-                IntArithOp::Sub => fn_ctx
-                    .builder
-                    .build_int_compare(IntPredicate::SLT, rhs_v, zero, "sat_sub_negative")
-                    .llvm_ctx("saturating sub sign compare")?,
-                IntArithOp::Mul => {
-                    let lhs_neg = fn_ctx
-                        .builder
-                        .build_int_compare(IntPredicate::SLT, lhs_v, zero, "sat_mul_lhs_neg")
-                        .llvm_ctx("saturating mul lhs sign compare")?;
-                    let rhs_neg = fn_ctx
-                        .builder
-                        .build_int_compare(IntPredicate::SLT, rhs_v, zero, "sat_mul_rhs_neg")
-                        .llvm_ctx("saturating mul rhs sign compare")?;
-                    fn_ctx
-                        .builder
-                        .build_int_compare(IntPredicate::EQ, lhs_neg, rhs_neg, "sat_mul_same_sign")
-                        .llvm_ctx("saturating mul same-sign compare")?
-                }
-            };
-            Ok(fn_ctx
-                .builder
-                .build_select(choose_max, max, min, "sat_signed_bound")
-                .llvm_ctx("saturating bound select")?
-                .into_int_value())
-        }
     }
 }
 
@@ -12932,222 +12854,6 @@ fn lower_numeric_cast(
         .builder
         .build_store(dest_ptr, casted)
         .llvm_ctx("numeric cast store")?;
-    Ok(())
-}
-
-/// Saturating integer-to-integer width conversion.
-///
-/// Both `from_ty` and `to_ty` must be checker-admitted integers.
-/// The source value is clamped to `[to_ty::MIN, to_ty::MAX]` before
-/// the narrowing (or widening) cast.
-///
-/// Algorithm overview:
-///
-/// - **Widening** (`src_bits < dest_bits`): every source value is already
-///   representable in the wider target — no clamping needed.  Just extend
-///   (sext for signed sources, zext for unsigned sources).
-///
-/// - **Same-width** (`src_bits == dest_bits`) or **Narrowing**
-///   (`src_bits > dest_bits`): the target bounds fit in the source-width
-///   type, so work entirely in the source integer type.  Emit two
-///   `cmp + select` pairs to clamp, then truncate / bit-cast as needed.
-///
-///   Bounds computation:
-///   - Signed target  k bits: MIN = -(2^(k-1)),  MAX = 2^(k-1) - 1
-///   - Unsigned target k bits: MIN = 0,           MAX = 2^k - 1
-///
-///   Represented as constants in the *source* integer type; they fit
-///   because src_bits >= dest_bits in this path.
-///
-///   Comparisons use the *source* signedness (we are asking "does the
-///   source value fall outside the target range?").
-fn lower_saturating_width_cast(
-    fn_ctx: &FnCtx<'_, '_>,
-    dest: Place,
-    src: Place,
-    from_ty: &ResolvedTy,
-    to_ty: &ResolvedTy,
-) -> CodegenResult<()> {
-    if !from_ty.is_integer() || !to_ty.is_integer() {
-        return Err(CodegenError::FailClosed(format!(
-            "SaturatingWidthCast requires integer types; got {} → {}",
-            from_ty.user_facing(),
-            to_ty.user_facing()
-        )));
-    }
-
-    let (src_ptr, src_storage) = place_pointer(fn_ctx, src)?;
-    let (dest_ptr, dest_storage) = place_pointer(fn_ctx, dest)?;
-    let src_int_ty = expect_int_type(src_storage, "saturating width cast source")?;
-    let dest_int_ty = expect_int_type(dest_storage, "saturating width cast dest")?;
-
-    let expected_src = primitive_to_llvm(fn_ctx.ctx, from_ty)?;
-    let expected_dest = primitive_to_llvm(fn_ctx.ctx, to_ty)?;
-    if src_storage != expected_src {
-        return Err(CodegenError::FailClosed(format!(
-            "SaturatingWidthCast source storage {src_storage:?} disagrees with from_ty {}",
-            from_ty.user_facing()
-        )));
-    }
-    if dest_storage != expected_dest {
-        return Err(CodegenError::FailClosed(format!(
-            "SaturatingWidthCast dest storage {dest_storage:?} disagrees with to_ty {}",
-            to_ty.user_facing()
-        )));
-    }
-
-    let src_v = fn_ctx
-        .builder
-        .build_load(src_int_ty, src_ptr, "sat_wcast_src")
-        .llvm_ctx("saturating width cast source load")?
-        .into_int_value();
-
-    let src_bits = src_int_ty.get_bit_width();
-    let dest_bits = dest_int_ty.get_bit_width();
-    let to_signed = to_ty.is_signed_integer();
-    let from_signed = from_ty.is_signed_integer();
-
-    let result: BasicValueEnum<'_> = if src_bits < dest_bits {
-        // Widening: every source value fits in the target — no clamping.
-        // Extend using source signedness (sext for signed, zext for unsigned).
-        if from_signed {
-            fn_ctx
-                .builder
-                .build_int_s_extend(src_v, dest_int_ty, "sat_wcast_sext")
-                .llvm_ctx("saturating width cast sign extend")?
-                .into()
-        } else {
-            fn_ctx
-                .builder
-                .build_int_z_extend(src_v, dest_int_ty, "sat_wcast_zext")
-                .llvm_ctx("saturating width cast zero extend")?
-                .into()
-        }
-    } else if src_bits == dest_bits && from_signed != to_signed {
-        // Same-width sign change: the target range is NOT a subset of the source
-        // range when interpreted in the source signedness, so we cannot compute
-        // clamping bounds in the source type.  Each direction needs only ONE clamp:
-        //
-        //   signed → unsigned (e.g. i32 → u32):
-        //     Target range [0, 2^k - 1].  Every non-negative source value fits
-        //     exactly; negative source values clamp to 0.  No upper clamp is
-        //     needed because 2^k - 1 is not representable as a signed k-bit value.
-        //     Emit: max(src, 0) via SGT + select, then let the store bitcast.
-        //
-        //   unsigned → signed (e.g. u32 → i32):
-        //     Target range [-(2^(k-1)), 2^(k-1)-1].  Source is always >= 0 so no
-        //     lower clamp is needed.  Clamp high at 2^(k-1)-1 using UGT so the
-        //     comparison sees the unsigned magnitude correctly.
-        //     Emit: min(src, 2^(k-1)-1) via UGT + select, then let the store bitcast.
-        if from_signed {
-            // signed → unsigned: clamp low at 0 only.
-            let zero = src_int_ty.const_zero();
-            let below_zero = fn_ctx
-                .builder
-                .build_int_compare(IntPredicate::SLT, src_v, zero, "sat_swcast_below_zero")
-                .llvm_ctx("saturating same-width sign cast below-zero compare")?;
-            fn_ctx
-                .builder
-                .build_select(below_zero, zero, src_v, "sat_swcast_clamped")
-                .llvm_ctx("saturating same-width sign cast clamp low select")?
-                .into_int_value()
-                .into()
-        } else {
-            // unsigned → signed: clamp high at 2^(k-1)-1 only.
-            // The constant is representable as an unsigned k-bit value.
-            let signed_max_val: u64 = (1_u64 << (dest_bits - 1)) - 1;
-            let signed_max = src_int_ty.const_int(signed_max_val, false);
-            let above_max = fn_ctx
-                .builder
-                .build_int_compare(IntPredicate::UGT, src_v, signed_max, "sat_swcast_above_max")
-                .llvm_ctx("saturating same-width sign cast above-max compare")?;
-            fn_ctx
-                .builder
-                .build_select(above_max, signed_max, src_v, "sat_swcast_clamped")
-                .llvm_ctx("saturating same-width sign cast clamp high select")?
-                .into_int_value()
-                .into()
-        }
-    } else {
-        // Narrowing (src_bits > dest_bits): clamp in the source integer type, then
-        // truncate.  The target bounds always fit in the source type here because
-        // src_bits > dest_bits.
-        let (min_const, max_const) = if to_signed {
-            // Signed target k bits: MIN = -(2^(k-1)), MAX = 2^(k-1) - 1
-            let min_val: i64 = -1_i64 << (dest_bits - 1);
-            let max_val: i64 = (1_i64 << (dest_bits - 1)) - 1;
-            (
-                src_int_ty.const_int(min_val as u64, true),
-                src_int_ty.const_int(max_val as u64, true),
-            )
-        } else {
-            // Unsigned target k bits: MIN = 0, MAX = 2^k - 1
-            // dest_bits < src_bits <= 64, so the shift is safe.
-            let max_val: u64 = (1_u64 << dest_bits) - 1;
-            (
-                src_int_ty.const_zero(),
-                src_int_ty.const_int(max_val, false),
-            )
-        };
-
-        // Clamp below: if src < min → use min.
-        // Compare using source signedness.
-        let below_min = if from_signed {
-            fn_ctx
-                .builder
-                .build_int_compare(IntPredicate::SLT, src_v, min_const, "sat_wcast_below_min")
-                .llvm_ctx("saturating width cast below-min compare")?
-        } else {
-            fn_ctx
-                .builder
-                .build_int_compare(IntPredicate::ULT, src_v, min_const, "sat_wcast_below_min")
-                .llvm_ctx("saturating width cast below-min compare")?
-        };
-        let clamped_low = fn_ctx
-            .builder
-            .build_select(below_min, min_const, src_v, "sat_wcast_clamp_low")
-            .llvm_ctx("saturating width cast clamp low select")?
-            .into_int_value();
-
-        // Clamp above: if clamped_low > max → use max.
-        let above_max = if from_signed {
-            fn_ctx
-                .builder
-                .build_int_compare(
-                    IntPredicate::SGT,
-                    clamped_low,
-                    max_const,
-                    "sat_wcast_above_max",
-                )
-                .llvm_ctx("saturating width cast above-max compare")?
-        } else {
-            fn_ctx
-                .builder
-                .build_int_compare(
-                    IntPredicate::UGT,
-                    clamped_low,
-                    max_const,
-                    "sat_wcast_above_max",
-                )
-                .llvm_ctx("saturating width cast above-max compare")?
-        };
-        let clamped = fn_ctx
-            .builder
-            .build_select(above_max, max_const, clamped_low, "sat_wcast_clamped")
-            .llvm_ctx("saturating width cast clamp high select")?
-            .into_int_value();
-
-        fn_ctx
-            .builder
-            .build_int_truncate(clamped, dest_int_ty, "sat_wcast_trunc")
-            .llvm_ctx("saturating width cast truncate")?
-            .into()
-    };
-
-    fn_ctx
-        .builder
-        .build_store(dest_ptr, result)
-        .llvm_ctx("saturating width cast store")?;
     Ok(())
 }
 
@@ -14087,7 +13793,7 @@ fn lower_instruction(
                     "IntArithCheckedOption Some payload type must match operand type".into(),
                 ));
             }
-            let intrinsic_name = overflow_intrinsic_name(*op, *signed);
+            let intrinsic_name = crate::arith::overflow_intrinsic_name(*op, *signed);
             let intrinsic = Intrinsic::find(intrinsic_name).ok_or_else(|| {
                 llvm_err!("with-overflow intrinsic `{intrinsic_name}` not found in LLVM build")
             })?;
@@ -14189,7 +13895,7 @@ fn lower_instruction(
             // `.with.overflow` + select path is kept for multiplication.
             let final_v = match op {
                 IntArithOp::Add | IntArithOp::Sub => {
-                    let intrinsic_name = saturating_intrinsic_name(*op, *signed);
+                    let intrinsic_name = crate::arith::saturating_intrinsic_name(*op, *signed);
                     let intrinsic = Intrinsic::find(intrinsic_name).ok_or_else(|| {
                         llvm_err!("sat intrinsic `{intrinsic_name}` not found in LLVM build")
                     })?;
@@ -14217,7 +13923,7 @@ fn lower_instruction(
                 IntArithOp::Mul => {
                     // `llvm.smul.sat` does not exist in LLVM; keep the
                     // with.overflow + saturating_bound + select sequence.
-                    let intrinsic_name = overflow_intrinsic_name(*op, *signed);
+                    let intrinsic_name = crate::arith::overflow_intrinsic_name(*op, *signed);
                     let intrinsic = Intrinsic::find(intrinsic_name).ok_or_else(|| {
                         llvm_err!(
                             "with-overflow intrinsic `{intrinsic_name}` not found in LLVM build"
@@ -14253,7 +13959,9 @@ fn lower_instruction(
                         .build_extract_value(agg, 1, "saturating_mul_overflow")
                         .llvm_ctx("saturating mul overflow extract")?
                         .into_int_value();
-                    let bound = saturating_bound(fn_ctx, *op, *signed, lhs_int, lhs_v, rhs_v)?;
+                    let bound = crate::arith::saturating_bound(
+                        fn_ctx, *op, *signed, lhs_int, lhs_v, rhs_v,
+                    )?;
                     fn_ctx
                         .builder
                         .build_select(of_bit, bound, result_v, "saturating_mul_select")
@@ -14730,7 +14438,7 @@ fn lower_instruction(
             from_ty,
             to_ty,
         } => {
-            lower_saturating_width_cast(fn_ctx, *dest, *src, from_ty, to_ty)?;
+            crate::arith::lower_saturating_width_cast(fn_ctx, *dest, *src, from_ty, to_ty)?;
             let _ = ctx;
         }
         Instr::Move { dest, src } => {
@@ -40276,158 +39984,6 @@ fn emit_wasm_main_export_wrapper<'ctx>(
     Ok(())
 }
 
-/// The memory-intrinsic floor (`mem.*`, W5.005 / F1b) for which codegen
-/// synthesizes a trampoline body directly from the catalog id threaded on
-/// `RawMirFunction::intrinsic_id`.
-///
-/// This enum is the central fail-closed authority (D343): an `intrinsic_id`
-/// that does not parse to one of these variants is a hard `CodegenError`, not
-/// a silent empty-body no-op. The catalog (`stdlib_catalog`) and the
-/// `std.mem` floor surface are the source of truth for the id strings; this
-/// `match` is the codegen-side counterpart and must stay in lockstep with
-/// them (the `unknown_floor_intrinsic_id_fails_closed` test guards the gap).
-#[derive(Debug, Clone, Copy, PartialEq, Eq)]
-enum FloorIntrinsic {
-    Alloc,
-    Realloc,
-    Dealloc,
-    PtrOffset,
-    PtrCopy,
-}
-
-impl FloorIntrinsic {
-    fn from_catalog_id(id: &str) -> Option<Self> {
-        match id {
-            "mem.alloc" => Some(Self::Alloc),
-            "mem.realloc" => Some(Self::Realloc),
-            "mem.dealloc" => Some(Self::Dealloc),
-            "mem.ptr_offset" => Some(Self::PtrOffset),
-            "mem.ptr_copy" => Some(Self::PtrCopy),
-            _ => None,
-        }
-    }
-}
-
-/// Synthesize the trampoline body for a `#[intrinsic("mem.*")]` memory-floor
-/// function (W5.005 / F1b, Decision 4 Option A).
-///
-/// The MIR `blocks` of these functions are a bodyless placeholder — the
-/// source declaration is `fn alloc(..) -> *mut u8 {}`. Codegen discards that
-/// placeholder and emits the real lowering here, driven by the typed
-/// `intrinsic` id threaded from HIR. Calls to `mem$alloc` (etc.) dispatch
-/// normally to this defined function; the body is the substrate.
-///
-/// Pointer args/returns are opaque `ptr`; the integer args are u64-width
-/// (`i64` in LLVM). Byte-level monomorphic (A612): `ptr_offset`/`ptr_copy`
-/// count raw bytes — the caller has already multiplied by the descriptor
-/// size, so no element-type scaling happens here.
-///
-/// Fail-closed (D343): the caller rejects an unrecognised id before reaching
-/// this function, so every arm maps to a concrete lowering — there is no
-/// silent empty-body path. A param-arity mismatch between the stub signature
-/// and the lowering is also a hard error.
-fn emit_floor_intrinsic_body<'ctx>(
-    ctx: &'ctx Context,
-    llvm_mod: &LlvmModule<'ctx>,
-    llvm_fn: FunctionValue<'ctx>,
-    func: &RawMirFunction,
-    intrinsic: FloorIntrinsic,
-) -> CodegenResult<()> {
-    let builder = ctx.create_builder();
-    let entry = ctx.append_basic_block(llvm_fn, "entry");
-    builder.position_at_end(entry);
-    let mut decls = RuntimeDeclMap::new();
-    let i8_ty = ctx.i8_type();
-
-    let param = |i: u32| -> CodegenResult<BasicValueEnum<'ctx>> {
-        llvm_fn.get_nth_param(i).ok_or_else(|| {
-            CodegenError::FailClosed(format!(
-                "floor intrinsic `{}` ({intrinsic:?}) has no LLVM parameter at index {i}; \
-                 the `std.mem` stub signature and the codegen lowering disagree",
-                func.name
-            ))
-        })
-    };
-
-    match intrinsic {
-        FloorIntrinsic::Alloc => {
-            let size = param(0)?;
-            let align = param(1)?;
-            let f = intern_runtime_decl(ctx, llvm_mod, &mut decls, "hew_alloc")?;
-            let call = builder
-                .build_call(f, &[size.into(), align.into()], "mem_alloc")
-                .llvm_ctx("mem.alloc hew_alloc call")?;
-            let ret = call.try_as_basic_value().basic().ok_or_else(|| {
-                CodegenError::FailClosed("hew_alloc returned void; expected ptr".into())
-            })?;
-            builder.build_return(Some(&ret)).llvm_ctx("mem.alloc ret")?;
-        }
-        FloorIntrinsic::Realloc => {
-            let ptr = param(0)?;
-            let old_size = param(1)?;
-            let new_size = param(2)?;
-            let align = param(3)?;
-            let f = intern_runtime_decl(ctx, llvm_mod, &mut decls, "hew_realloc")?;
-            let call = builder
-                .build_call(
-                    f,
-                    &[ptr.into(), old_size.into(), new_size.into(), align.into()],
-                    "mem_realloc",
-                )
-                .llvm_ctx("mem.realloc hew_realloc call")?;
-            let ret = call.try_as_basic_value().basic().ok_or_else(|| {
-                CodegenError::FailClosed("hew_realloc returned void; expected ptr".into())
-            })?;
-            builder
-                .build_return(Some(&ret))
-                .llvm_ctx("mem.realloc ret")?;
-        }
-        FloorIntrinsic::Dealloc => {
-            let ptr = param(0)?;
-            let size = param(1)?;
-            let align = param(2)?;
-            let f = intern_runtime_decl(ctx, llvm_mod, &mut decls, "hew_dealloc")?;
-            builder
-                .build_call(f, &[ptr.into(), size.into(), align.into()], "mem_dealloc")
-                .llvm_ctx("mem.dealloc hew_dealloc call")?;
-            builder
-                .build_return(Some(&i8_ty.const_zero()))
-                .llvm_ctx("mem.dealloc ret")?;
-        }
-        FloorIntrinsic::PtrOffset => {
-            let base = param(0)?.into_pointer_value();
-            let byte_offset = param(1)?.into_int_value();
-            // SAFETY: i8 in-bounds GEP by a raw byte count — this primitive IS
-            // Rust's `<*mut T>::add` (inbounds). The result must stay within the
-            // same allocation `base` points into (one-past-end permitted);
-            // anything else is UB. Bounds are the caller's obligation (A605 —
-            // unsafe primitives; the only callers are compiler-authored
-            // containers). Cross-allocation/wrapping offset is a separate future
-            // primitive, intentionally not provided here (YAGNI).
-            let gep = unsafe {
-                builder
-                    .build_in_bounds_gep(i8_ty, base, &[byte_offset], "mem_ptr_offset")
-                    .llvm_ctx("mem.ptr_offset i8 gep")?
-            };
-            builder
-                .build_return(Some(&gep))
-                .llvm_ctx("mem.ptr_offset ret")?;
-        }
-        FloorIntrinsic::PtrCopy => {
-            let dst = param(0)?.into_pointer_value();
-            let src = param(1)?.into_pointer_value();
-            let byte_count = param(2)?.into_int_value();
-            builder
-                .build_memcpy(dst, 1, src, 1, byte_count)
-                .llvm_ctx("mem.ptr_copy memcpy")?;
-            builder
-                .build_return(Some(&i8_ty.const_zero()))
-                .llvm_ctx("mem.ptr_copy ret")?;
-        }
-    }
-    Ok(())
-}
-
 fn function_needs_closure_call_fallback_context(func: &RawMirFunction) -> bool {
     !func.call_conv.carries_execution_context()
         && func.blocks.iter().any(|block| {
@@ -40616,15 +40172,16 @@ fn lower_function<'ctx>(
     // Fail-closed: an id codegen does not recognise is a hard error, never a
     // silent empty-body no-op (the historic fail-OPEN bug).
     if let Some(intrinsic_id) = &func.intrinsic_id {
-        let intrinsic = FloorIntrinsic::from_catalog_id(intrinsic_id).ok_or_else(|| {
-            CodegenError::FailClosed(format!(
-                "function `{}` carries unknown floor intrinsic id {intrinsic_id:?}; \
+        let intrinsic =
+            crate::arith::FloorIntrinsic::from_catalog_id(intrinsic_id).ok_or_else(|| {
+                CodegenError::FailClosed(format!(
+                    "function `{}` carries unknown floor intrinsic id {intrinsic_id:?}; \
                  codegen has no lowering for it — extend `FloorIntrinsic` or leave \
                  the declaration fail-closed (it must never emit an empty body)",
-                func.name
-            ))
-        })?;
-        return emit_floor_intrinsic_body(ctx, llvm_mod, llvm_fn, func, intrinsic);
+                    func.name
+                ))
+            })?;
+        return crate::arith::emit_floor_intrinsic_body(ctx, llvm_mod, llvm_fn, func, intrinsic);
     }
     let context_marker_findings = validate_context_markers_for_codegen(func);
     if !context_marker_findings.is_empty() {
